challenge2/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
pandas
pillow
matplotlib
flask_migrate
numpy
scipy